% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{adjust_discards}
\alias{adjust_discards}
\title{Adjust the input data for reported rates of discarding.}
\usage{
adjust_discards(dat, discards, regstr, runsp)
}
\arguments{
\item{dat}{Input dataset}

\item{discards}{Data frame containing the discards.}

\item{regstr}{The current regional structure.}

\item{runsp}{The species of interest.}
}
\value{
Modified dataset.
}
\description{
The function takes discard rates by flag, regional structure, region and year, and divides the catch by 1 / (1 - discard rate).
}
