% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{aggregate_by_trip}
\alias{aggregate_by_trip}
\title{Aggregate by trip.}
\usage{
aggregate_by_trip(dat, flds)
}
\arguments{
\item{dat}{The input dataset, at the set level.}

\item{flds}{The fields to aggregate by summation.}
}
\value{
The aggregated dataset.
}
\description{
The function aggregates the input dataset by trip using the variable TRIP_NUM.
}
