% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate the input dataset by yearqtr, latlong, and hbf.}
\usage{
aggregate_data(dat, sp)
}
\arguments{
\item{dat}{The input dataset, at the set level.}

\item{sp}{The species code of the species of interest.}
}
\value{
taball, the aggregated dataset.
}
\description{
Aggregate the input dataset by yearqtr, latlong, and hbf.
}
