% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{boxplots_PCA}
\alias{boxplots_PCA}
\title{Boxplots of PCA by set characteristics.}
\usage{
boxplots_PCA(dat, nPCA = 3, ti = "", dohbf = T, lat5 = F,
  regtype = "regY", r = r, saveplot = TRUE, savedir = ".")
}
\arguments{
\item{dat}{Input dataset}

\item{nPCA}{Number of PCA variables to plot.}

\item{ti}{Plot main title.}

\item{dohbf}{Boxplot PCA by HBF.}

\item{lat5}{Use 'lat' variable if FALSE, otherwise 'lat5'.}

\item{regtype}{Deprecated, not used.}

\item{r}{Deprecated, not used.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}

\item{savedir}{The directory to save the plot.}
}
\description{
This function plots beanplots of set characteristics by cluster.
}
