% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{boxplots_spCL_comp}
\alias{boxplots_spCL_comp}
\title{Boxplots of species composition by cluster.}
\usage{
boxplots_spCL_comp(
  dat,
  cl = "kmeans",
  ti = "",
  outL = T,
  nsp = 13,
  regtype = regtype,
  r = r,
  allsp,
  saveplot = TRUE,
  savedir = "."
)
}
\arguments{
\item{dat}{Input dataset}

\item{cl}{Cluster type to select from dataset.}

\item{ti}{Plot main title.}

\item{outL}{Include outliers if TRUE.}

\item{nsp}{Number of species used in clustering.}

\item{regtype}{Deprecated, not used.}

\item{r}{Deprecated, not used.}

\item{allsp}{Vector of species to use in the cluster plots.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}

\item{savedir}{The directory to save the plot.}
}
\description{
This function plots boxplots of species composition by cluster.
}
