% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{boxplots_spTPCA}
\alias{boxplots_spTPCA}
\title{Boxplots of species composition by principal component.}
\usage{
boxplots_spTPCA(
  dat,
  nPCA = 6,
  ti = "",
  outL = T,
  nsp = 13,
  regtype = "regY",
  r = r,
  allsp,
  saveplot = TRUE,
  savedir = "."
)
}
\arguments{
\item{dat}{Input dataset}

\item{nPCA}{Number of principal components to plot.}

\item{ti}{Plot main title.}

\item{outL}{Include outliers if TRUE.}

\item{nsp}{Number of species used in clustering.}

\item{regtype}{Deprecated, not used.}

\item{r}{Deprecated, not used.}

\item{allsp}{Vector of species to use in the PCA plots.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}

\item{savedir}{The directory to save the plot.}
}
\description{
This function plots boxplots of species composition by principal component.
}
