% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{clust_PCA_run}
\alias{clust_PCA_run}
\title{Generate clusters, PCs, and associated figures}
\usage{
clust_PCA_run(r, ddd, allsp, allabs, regtype = "regY", ncl, plotPCA = T,
  clustid = "tripidmon", allclust = F, flag = "JP", dbh = TRUE,
  cllist = NA, fnhead = "", ll5 = F, covarnames = c("yrqtr", "latlong",
  "hooks", "hbf", "vessid", "callsign", "Total", "lat", "lon", "lat5", "lon5",
  "moon", "op_yr", "op_mon"))
}
\arguments{
\item{r}{The region number.}

\item{ddd}{The dataset to be clustered.}

\item{allsp}{Vector of species codes to use in catch composition.}

\item{allabs}{Vector of variable names.}

\item{regtype}{Label to identify which regional structure to use for region numbers.}

\item{ncl}{Number of clusters to use.}

\item{plotPCA}{Generate PCA plots if TRUE.}

\item{clustid}{Name of variable to aggregate across for trip-level clustering.}

\item{allclust}{Generate plots for all clustering methods if TRUE.}

\item{flag}{Fleet code to use in figure titles.}

\item{dbh}{If TRUE, create plots of HBF by cluster.}

\item{cllist}{cllist}

\item{fnhead}{File name head}

\item{ll5}{If TRUE, ll5 variables are used instead of 1 degree versions}

\item{covarnames}{Names of covariates to return with dataset.}
}
\value{
The original dataset with added columns for the clusters and PCAs;
}
\description{
Top level function that calls functions to make clusters and principal components (PCA), and to plot the results, and associates clusters and PCs with each row of the original dataset.
}
