% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{combine_delta}
\alias{combine_delta}
\title{Combine binomial and lognormal indices.}
\usage{
combine_delta(fnamedelta, fnamepos, fname, runsp, runreg)
}
\arguments{
\item{fnamedelta}{Name of the delta object to load.}

\item{fnamepos}{Name of the lognormal positive object to load.}

\item{fname}{File name used for saving outputs.}

\item{runsp}{Species of interest.}

\item{runreg}{The model region, used in output file names.}
}
\description{
The function combines binomial and lognormal positive indices to generate an index of relative abundance.
}
