% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{combine_delta_xl}
\alias{combine_delta_xl}
\title{Combine binomial and lognormal indices, importing from excel.}
\usage{
combine_delta_xl(fnamedelta, fnamepos, runsp, runreg)
}
\arguments{
\item{fnamedelta}{Name of the delta file to load.}

\item{fnamepos}{Name of the lognormal positive file to load.}

\item{runsp}{Species of interest.}

\item{runreg}{The model region, used in output file names.}
}
\description{
The function combines binomial and lognormal positive indices to generate an index of relative abundance.
}
