% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{countzero}
\alias{countzero}
\title{count proportion of zeroes in a vector.}
\usage{
countzero(a)
}
\arguments{
\item{a}{Input vector.}
}
\value{
Proportion of zeroes.
}
\description{
A helper function to calulate the proportion of zero values in a vector.
}
