% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean}
\alias{dataclean}
\title{Clean data.}
\usage{
dataclean(dat, checktg = F, allHBF = F, splist)
}
\arguments{
\item{dat}{Input dataset}

\item{checktg}{If TRUE, remove effort identified as not targetng tuna. Not generally used.}

\item{allHBF}{Deprecated.}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Japanese data in the WCPO.
}
