% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_BR}
\alias{dataclean_BR}
\title{Clean Brazilian data.}
\usage{
dataclean_BR(dat, yearlim = 2018, splist)
}
\arguments{
\item{dat}{Input dataset}

\item{yearlim}{Data is included only from before this year.}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data.
}
