% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_CNIO}
\alias{dataclean_CNIO}
\title{Clean data.}
\usage{
dataclean_CNIO(
  dat,
  checktg = F,
  allHBF = F,
  splist = c("bft", "sbt", "alb", "bet", "yft", "swo", "mls", "bum", "blm", "sas",
    "sha")
)
}
\arguments{
\item{dat}{Input dataset}

\item{checktg}{If TRUE, remove effort identified as not targetng tuna. Not generally used.}

\item{allHBF}{Deprecated.}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Japanese data in the IO.
}
