% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_JP_EPO}
\alias{dataclean_JP_EPO}
\title{Clean data.}
\usage{
dataclean_JP_EPO(
  dat,
  checktg = F,
  splist = c("alb", "bet", "yft", "swo", "mls", "bum", "blm", "sas", "sha", "sfa", "ssp")
)
}
\arguments{
\item{dat}{Input dataset}

\item{checktg}{If TRUE, remove effort identified as not targetng tuna. Not generally used.}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Japanese data in the IO.
}
