% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_KR}
\alias{dataclean_KR}
\title{Clean data.}
\usage{
dataclean_KR(dat, yearlim = 2016)
}
\arguments{
\item{dat}{Input dataset}

\item{yearlim}{Upper boundary for time period.}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Korean data in the IO.
}
