% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_SY}
\alias{dataclean_SY}
\title{Clean SY data.}
\usage{
dataclean_SY(dat1, rmssp = F, splist = c("alb", "bet", "yft", "ott", "swo",
  "mls", "bum", "blm", "otb", "skj", "sha", "oth", "sbt"))
}
\arguments{
\item{dat1}{Input dataset}

\item{rmssp}{If TRUE, remove sets that report catch of only one species.}

\item{splist}{List of species codes.}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Taiwanese data in the IO.
}
