% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_TW}
\alias{dataclean_TW}
\title{Clean data.}
\usage{
dataclean_TW(dat1, rmssp = F)
}
\arguments{
\item{dat1}{Input dataset}

\item{rmssp}{If TRUE, remove sets that report catch of only one species.}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Taiwanese data in the IO.
}
