% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_TW_EPO}
\alias{dataclean_TW_EPO}
\title{Clean TW data.}
\usage{
dataclean_TW_EPO(dat1, rmssp = F, splist = c("alb", "bet", "yft",
  "ott", "swo", "mls", "bum", "blm", "otb", "skj", "sha", "oth"))
}
\arguments{
\item{dat1}{Input dataset}

\item{rmssp}{If TRUE, remove sets that report catch of only one species.}

\item{splist}{List of species codes.}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Taiwanese data in the IO.
}
