% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_USAO}
\alias{dataclean_USAO}
\title{Clean US data.}
\usage{
dataclean_USAO(dat, splist = c("bft", "alb", "bet", "yft", "swo", "mls",
  "bum", "bsh", "sma", "por"))
}
\arguments{
\item{dat}{Input dataset}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data.
}
