% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_BRA}
\alias{dataprep_BRA}
\title{Prepare Brazilian longline data.}
\usage{
dataprep_BRA(dat, splist)
}
\arguments{
\item{dat}{Input dataset}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function prepares Brazilian longline data.
Developed for AO analyses by Rodrigo Sant'Ana.
}
