% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_JP}
\alias{dataprep_JP}
\title{Prepare Japanese longline data.}
\usage{
dataprep_JP(
  dat,
  alldat = T,
  region = "IO",
  splist = c("bft", "sbt", "alb", "bet", "yft", "swo", "mls", "bum", "blm", "sas", "sha")
)
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{If FALSE, removes vessels without vessel ID.}

\item{region}{Set the region to AO or IO.}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function prepares Japanese longline data for IO and AO analyses.
}
