% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_JP_EPO}
\alias{dataprep_JP_EPO}
\title{Prepare Japanese longline data for the eastern and western Pacific, IATTC 2019.}
\usage{
dataprep_JP_EPO(dat, alldat = T, splist = c("alb", "bet", "yft", "swo",
  "mls", "bum", "blm", "sas", "sha", "sai", "spf"))
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{If FALSE, removes vessels without vessel ID.}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function prepares Japanese longline data for EPO (mainly) but also some WCPO analyses. Data are already partly preprared.
}
