% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_JPIO}
\alias{dataprep_JPIO}
\title{Prepare Japanese longline data.}
\usage{
dataprep_JPIO(dat, alldat = T)
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{If FALSE, removes vessels without vessel ID.}
}
\value{
Modified dataset.
}
\description{
The function prepares Japanese longline data for IO analyses.
}
