% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_KRSBT}
\alias{dataprep_KRSBT}
\title{Prepare Korean SBT data.}
\usage{
dataprep_KRSBT(dat, alldat = F, splist)
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{Deprecated.}
}
\value{
Modified dataset.
}
\description{
The function prepares the Korean SBT data.
}
