% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_SY}
\alias{dataprep_SY}
\title{Prepare Seychelles longline data.}
\usage{
dataprep_SY(dat, region, splist)
}
\arguments{
\item{dat}{Input dataset}

\item{region}{IO or AO.}

\item{splist}{Define the species in the dataset}

\item{alldat}{Not used.}
}
\value{
Modified dataset.
}
\description{
The function prepares Seychelles longline data for IO analyses.
}
