% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_TW_EPO}
\alias{dataprep_TW_EPO}
\title{Prepare Taiwanese longline data for the EPO.}
\usage{
dataprep_TW_EPO(dat, alldat = F, region = "EPO", splist = c("alb",
  "bet", "yft", "ott", "swo", "mls", "bum", "blm", "otb", "skj", "skx",
  "oth"))
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{Not used.}

\item{region}{Only EPO at the moment.}

\item{splist}{Define the species in the dataset}
}
\value{
Modified dataset.
}
\description{
The function prepares Taiwanese longline data for EPO analyses.
}
