% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_US}
\alias{dataprep_US}
\title{Prepare US longline data.}
\usage{
dataprep_US(dat, splist)
}
\arguments{
\item{dat}{Input dataset}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function prepares US longline data. Developed for AO analyses.
}
