% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{defactor}
\alias{defactor}
\title{Turn a factor into a numeric.}
\usage{
defactor(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
Numeric vector.
}
\description{
A helper function to turn a vector of factors into numeric variables.
}
