% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{do_deltalog}
\alias{do_deltalog}
\title{Run delta lognormal models.}
\usage{
do_deltalog(
  dat,
  dohbf = F,
  addboat = F,
  addcl = addcl,
  nhbf = 3,
  runsp,
  fname,
  modlab,
  keepd = TRUE,
  dohook = TRUE
)
}
\arguments{
\item{dat}{Input dataset.}

\item{dohbf}{Include HBF in the models.}

\item{addboat}{Include vess in the models.}

\item{addcl}{Include clust in the models.}

\item{nhbf}{Number of knots in the hbf spline.}

\item{runsp}{Species of interest.}

\item{fname}{Filename for saving.}

\item{modlab}{Label for plots and filenames.}

\item{keepd}{If FALSE, delete the data from the model object before saving.}

\item{dohook}{Include hooks in the model on the RHS if TRUE.}
}
\description{
The function runs delta lognormal models.
}
