% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{do_deltalogx}
\alias{do_deltalogx}
\title{Run delta lognormal models with interactions.}
\usage{
do_deltalogx(dat, dohbf = F, addboat = F, addcl = addcl, nhbf = 3,
  runsp, fname, modlab, keepd = TRUE, lat5xqtr = T, lat5xyr = T,
  bcorr = FALSE, cell_areas)
}
\arguments{
\item{dat}{Input dataset.}

\item{dohbf}{Include HBF in the models.}

\item{addboat}{Include vess in the models.}

\item{addcl}{Include clust in the models.}

\item{nhbf}{Number of knots in the hbf spline.}

\item{runsp}{Species of interest.}

\item{fname}{Filename for saving.}

\item{modlab}{Label for plots and filenames.}

\item{keepd}{If FALSE, delete the data from the model object before saving.}

\item{lat5xqtr}{Include lat5 by qtr interaction in the model.}

\item{lat5xyr}{Include lat5 by yr interaction in the model.}

\item{bcorr}{Include lognormal bias correction.}

\item{cell_areas}{Cell areas to use when aggregating densities across space.}
}
\description{
The function runs delta lognormal glm models with interactions.
}
