% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{doplot_cpue}
\alias{doplot_cpue}
\title{Plot quarterly indices.}
\usage{
doplot_cpue(a, vartype, mdti, regstr, runreg)
}
\arguments{
\item{a}{The input data frame.}

\item{vartype}{Type of analysis, to use in the plot title.}

\item{mdti}{Period and type of run, to use in plot title.}

\item{regstr}{Name of regional structure, to use in file name.}

\item{runreg}{Region number, to use in file name.}
}
\description{
The function takes a data frame and plots it.
}
