% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{get.bin.coefs}
\alias{get.bin.coefs}
\title{Extract yrqtr coefficients from a binomial glm.}
\usage{
get.bin.coefs(model, nyrs = NA, dat = NA)
}
\arguments{
\item{model}{The model result object from the binomial glm.}

\item{nyrs}{Deprecated, not used.}

\item{dat}{Deprecated, not used.}
}
\value{
Yrqtr coefficients as proportions, with the same mean as the observed proportions.
}
\description{
The function extracts yrqtr coefficients from a binomial glm.
}
