% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{get.bincoefsx}
\alias{get.bincoefsx}
\title{Generate area-averaged yrqtr coefs from predicted binomial CPUE.}
\usage{
get.bincoefsx(nd, cell_areas)
}
\arguments{
\item{nd}{The predicted dataset by space and time.}

\item{cell_areas}{Object with cell areas to use in summing densities across areas.}
}
\value{
Yrqtr coefficients.
}
\description{
The function generates area-averaged yrqtr coefficients from the predicted dataset for a binomial glm with interactions.
}
