% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{get.coefs}
\alias{get.coefs}
\title{Extract yrqtr coefficients from a lognormal positive glm object.}
\usage{
get.coefs(model)
}
\arguments{
\item{model}{The summary object from the lognormal glm.}
}
\value{
Normalized yrqtr coefficients.
}
\description{
The function extracts yrqtr coefficients from a lognormal positive glm object.
}
