% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{get.coefs.summ}
\alias{get.coefs.summ}
\title{Extract yrqtr coefficients from a lognormal positive glm summary object.}
\usage{
get.coefs.summ(modelsumm, nyrs)
}
\arguments{
\item{modelsumm}{The summary object from the lognormal glm.}

\item{nyrs}{The number of yrqtr time variables to extract.}
}
\value{
Normalized yrqtr coefficients.
}
\description{
The function extracts yrqtr coefficients from a lognormal positive glm summary object.
}
