% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{get.coefsx}
\alias{get.coefsx}
\title{Generate area-averaged yrqtr coefs from predicted lognormal constant CPUE, with interactions.}
\usage{
get.coefsx(nd, mn, bcorr = F, cell_areas)
}
\arguments{
\item{nd}{The predicted dataset by space and time.}

\item{mn}{The mean value used as the constant.}

\item{bcorr}{Use lognormal bias correction if TRUE.}

\item{cell_areas}{Object with cell areas to use in summing densities across areas.}
}
\value{
Normalized yrqtr coefficients.
}
\description{
The function generates area-averaged yrqtr coefficients from the predicted dataset for a lognormal constant glm with interactions.
}
