% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{get.poscoefsx}
\alias{get.poscoefsx}
\title{Generate area-averaged yrqtr coefs from predicted lognormal positive CPUE.}
\usage{
get.poscoefsx(nd, bcorr = F, cell_areas)
}
\arguments{
\item{nd}{The predicted dataset by space and time.}

\item{bcorr}{Use lognormal bias correction if TRUE.}

\item{cell_areas}{Object with cell areas to use in summing densities across areas.}
}
\value{
Normalized yrqtr coefficients.
}
\description{
The function extracts yrqtr coefficients from the predicted dataset for a lognormal positive glm with interactions.
}
