% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{get.summ.coefs}
\alias{get.summ.coefs}
\title{Extract yrqtr coefficients from a lognormal positive glm summary object.}
\usage{
get.summ.coefs(model, nyrs)
}
\arguments{
\item{nyrs}{The number of yrqtr time variables to extract.}

\item{mod}{The summary object from the lognormal glm.}
}
\value{
Normalized yrqtr coefficients.
}
\description{
The function extracts yrqtr coefficients from a lognormal positive glm summary object. Same as get.coefs.summ().
}
