% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdat_functions.R
\name{get_base_newdat}
\alias{get_base_newdat}
\title{Make newdat object, generalised.}
\usage{
get_base_newdat(dat)
}
\arguments{
\item{dat}{Original dataset input to the glm.}
}
\value{
newdat.
}
\description{
Function to make newdat object with a generalised function applicable to all input datasets.
}
