% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_functions.R
\name{glm.coefs}
\alias{glm.coefs}
\title{Predict CPUE from a lognormal positive glm object.}
\usage{
glm.coefs(model, dat)
}
\arguments{
\item{model}{The fitted model object.}

\item{dat}{Dataset used in the glm.}
}
\value{
Predicted catch rates by yrqtr with CIs.
}
\description{
The function predicts CPUE lognormal positive glm object.
}
