% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{limit_vessels}
\alias{limit_vessels}
\title{Select vessels based on quarters fished.}
\usage{
limit_vessels(gdat, minqtrs, maxqtrs = 10000)
}
\arguments{
\item{gdat}{Input dataset}

\item{minqtrs}{Minimum number of quartrs fished by a vessel.}

\item{maxqtrs}{Maximum number of quartrs fished by a vessel, defaults to very high number.}
}
\value{
Modified dataset.
}
\description{
The function reduces the size of the dataset according to criteria about number of quarters fished by a vessel.
}
