% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ls.objects}
\alias{ls.objects}
\title{List objects in the workspace and their characteristics.}
\usage{
ls.objects(pos = 1, pattern, order.by, decreasing = FALSE,
  head = FALSE, n = 5)
}
\arguments{
\item{pos}{Search level.}

\item{pattern}{Pattern.}

\item{order.by}{Sort the output by this characteristic.}

\item{decreasing}{Sort descending if TRUE.}

\item{head}{Report the first n objects if head is TRUE.}

\item{n}{Report the first n objects if head is TRUE.}
}
\value{
Output to console.
}
\description{
List objects in the workspace and their characteristics.
}
