% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lsos}
\alias{lsos}
\title{List objects in the workspace and their characteristics.}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{Whatever.}

\item{n}{Report the first n objects if head is TRUE.}
}
\value{
Output to console.
}
\description{
List objects in the workspace and their characteristics.
}
