% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lu}
\alias{lu}
\alias{lu}
\title{Calulate number of unique values in a vector.}
\usage{
lu(x)

lu(x)
}
\arguments{
\item{x}{Input vector.}

\item{x}{Input vector.}
}
\value{
Number of unique values.

Number of unique values.
}
\description{
A helper function to calulate the number of unique values in a vector, length(unique()).

A helper function to calulate the number of unique values in a vector, length(unique()).
}
