% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{make_clid}
\alias{make_clid}
\title{Make various types of id for use in aggregating data before clustering.
The function makes multiple passes through the dataset.}
\usage{
make_clid(dat)
}
\arguments{
\item{dat}{The input dataset, at the set level.}
}
\value{
dat, the input dataset with the following variables added: latlong1, latlong2, clid, aggid, jnt_clid.
}
\description{
Make various types of id for use in aggregating data before clustering.
The function makes multiple passes through the dataset.
}
