% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{make_clusters}
\alias{make_clusters}
\title{Cluster dataset based on species composition.
Run cluster analysis at both set and trip levels, using three methods: hclust, clara, and kmeans.}
\usage{
make_clusters(setdat, spp, ncl = 5, titx = "", setclust = T,
  tripid = "tripid", fname = "", regtype = "regY", allsp)
}
\arguments{
\item{setdat}{The input dataset, at the set level.}

\item{spp}{The species codes to use for clustering.}

\item{ncl}{Number of clusters to use.}

\item{titx}{Title for plots.}

\item{setclust}{Do clustering at set level if TRUE.}

\item{tripid}{Name of variable used to aggreate for trip-level clustering.}

\item{fname}{File name header for saving plots.}

\item{regtype}{Region type code.}

\item{allsp}{All the species codes.}
}
\value{
A list of four objects: d, the normalised trip-level dataset; fit, the hclust results; clarax, the clara trip level results; setdat, all the clustering results, as additional columns added to the input dataset.
}
\description{
Cluster dataset based on species composition.
Run cluster analysis at both set and trip levels, using three methods: hclust, clara, and kmeans.
}
