% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_functions.R
\name{make_formula_IO}
\alias{make_formula_IO}
\title{Create formula object to use in the glm.}
\usage{
make_formula_IO(
  runsp,
  modtype,
  addboat,
  dohbf = T,
  splitboat = F,
  addcl = F,
  addpca = NA,
  nhbf = 6,
  dohook = TRUE
)
}
\arguments{
\item{runsp}{Species of interest.}

\item{modtype}{logn (lognormal constant), deltabin(binomial), deltapos (lognormal positive), qp (quasi-Poisson), propn (proportion, which uses runsp directly), negbin (negative binomial), or weibull (Weibull).}

\item{addboat}{Include vess if TRUE.}

\item{dohbf}{Include hbf if TRUE, with nhbf knots in spline.}

\item{splitboat}{Include splitvess if TRUE.}

\item{addcl}{Include clust if TRUE.}

\item{addpca}{Include PCA if TRUE.}

\item{nhbf}{Number of knots in hbf spline, if dohbf is TRUE.}

\item{dohook}{Include a spline on hooks if TRUE.}
}
\value{
Modified dataset.
}
\description{
This function creates a formula object for the glm, based on the input parameters. Developed for analyses in the IO.
}
