% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_functions.R
\name{make_formula_IOx}
\alias{make_formula_IOx}
\title{Create formula object to use in the glm, with interactions.}
\usage{
make_formula_IOx(runsp, modtype, addboat, dohbf = T, splitboat = F,
  addcl = F, addpca = NA, nhbf = 6, lat5xqtr = F, lat5xyr = F)
}
\arguments{
\item{runsp}{Species of interest.}

\item{modtype}{logn (lognormal constant), deltabin(binomial), deltapos (lognormal positive), qp (quasi-Poisson), propn (proportion, which uses runsp directly), negbin (negative binomial), or weibull (Weibull).}

\item{addboat}{Include vess if TRUE.}

\item{dohbf}{Include hbf if TRUE, with nhbf knots in spline.}

\item{splitboat}{Include splitvess if TRUE.}

\item{addcl}{Include clust if TRUE.}

\item{addpca}{Include PCA if TRUE.}

\item{nhbf}{Number of knots in hbf spline, if dohbf is TRUE.}

\item{lat5xqtr}{Include lat5:qtr if TRUE.}

\item{lat5xyr}{Include lat5:op_yr if TRUE.}
}
\value{
Modified dataset.
}
\description{
This function creates a formula object for the glm, based on the input parameters. Developed for analyses in the IO with interaction models.
}
