% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{make_index_windows}
\alias{make_index_windows}
\title{Set up output window for index plots}
\usage{
make_index_windows(wh = 12, ww = 12, wmf = c(2, 2))
}
\arguments{
\item{wh}{Window height}

\item{ww}{Window width}

\item{wmf}{Window mfrow parameter}
}
\description{
Set up output window for index plots
}
