% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{make_lbidmon}
\alias{make_lbidmon}
\title{Prepare the variable lbid_mon.}
\usage{
make_lbidmon(dat)
}
\arguments{
\item{dat}{Input dataset}
}
\value{
Modified dataset.
}
\description{
The function creates the variable lbid_mon which is assigned to Japanese data for clustering before vessel ids are available.
}
