% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdat_functions.R
\name{make_newdat_JP}
\alias{make_newdat_JP}
\title{Make newdat object and predict responses from the glm.}
\usage{
make_newdat_JP(model, datx)
}
\arguments{
\item{model}{GLM object to predict for.}

\item{datx}{Original dataset input to the glm.}
}
\value{
List object comprising newdat (new dataset), predresp (predicted responses) and predterms (individual components of the response).
}
\description{
Function to make newdat object and predict responses from the glm. Originally developed for Japanese data in the IO.
}
