% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdat_functions.R
\name{make_newdat_x}
\alias{make_newdat_x}
\title{Make newdat object.}
\usage{
make_newdat_x(mod, dat)
}
\arguments{
\item{mod}{GLM object to predict for.}

\item{dat}{Original dataset input to the glm.}
}
\value{
newdat.
}
\description{
Function to make newdat object and predict responses from the glm with time area interactions. Originally developed for joint analyses in the IO.
}
