% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_strat_wts_functions.R
\name{make_strat}
\alias{make_strat}
\title{Stratify dataset prior to sampling.}
\usage{
make_strat(dat)
}
\arguments{
\item{dat}{Input dataset}
}
\value{
A vector indicating the number of samples in the stratum of which the row is a member.
}
\description{
The function stratifies the dataset by grid cell and yrqtr, and returns a vector indicating the number of samples in the stratum of which the row is a member.
}
