% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{map_clusters}
\alias{map_clusters}
\title{Maps of cluster distribution.}
\usage{
map_clusters(
  ddd,
  cl = "hclustcl",
  ti = "",
  lat5 = F,
  regtype = "regY",
  ncl,
  r = r,
  xl = NA,
  yl = NA,
  saveplot = TRUE,
  savedir = "."
)
}
\arguments{
\item{ddd}{Input dataset}

\item{cl}{Cluster type to select from dataset.}

\item{ti}{Plot main title.}

\item{lat5}{If TRUE indicates that data are 5x5 rather than 1x1.}

\item{regtype}{Deprecated, not used.}

\item{ncl}{Number of clusters.}

\item{r}{Deprecated, not used.}

\item{xl}{Longitude range.}

\item{yl}{Latitude range.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}

\item{savedir}{Directory to save the plot in, defaulting to current.}
}
\description{
This function plots maps of cluster spatial distribution.
}
