% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{map_EPO}
\alias{map_EPO}
\title{Map of the Pacific Ocean with EPO boundaries.}
\usage{
map_EPO(new = FALSE, latlim = c(-40, 40), lonlim = c(140, 290))
}
\description{
Function to make a map of the Pacific Ocean, with regional boundaries.
}
