% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{mapPCA}
\alias{mapPCA}
\title{Maps of PCA distribution.}
\usage{
mapPCA(ddd, nPCA = 6, ti = "", lat5 = F, regtype = "regY", r = r,
  saveplot = TRUE, savedir = ".")
}
\arguments{
\item{ddd}{Input dataset}

\item{nPCA}{Number of Principal component plots.}

\item{ti}{Plot main title.}

\item{lat5}{If TRUE indicates that data are 5x5 rather than 1x1.}

\item{regtype}{Deprecated, not used.}

\item{r}{Deprecated, not used.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}

\item{savedir}{The directory to save the plot.}
}
\description{
This function plots maps of PCA spatial distribution.
}
