% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{mapTPCA}
\alias{mapTPCA}
\title{Maps of TCPA distribution.}
\usage{
mapTPCA(ddd, nPCA = 6, ti = "", lat5 = F, regtype = "regY",
  r = r)
}
\arguments{
\item{ddd}{Input dataset}

\item{nPCA}{Number of Principal component plots.}

\item{ti}{Plot main title.}

\item{lat5}{If TRUE indicates that data are 5x5 rather than 1x1.}

\item{regtype}{Deprecated, not used.}

\item{r}{Deprecated, not used.}
}
\description{
This function plots maps of TCPA distribution.
}
