% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_strat_wts_functions.R
\name{mk_wts}
\alias{mk_wts}
\title{Allocate statistical weights to each row.}
\usage{
mk_wts(dat, wttype, catch = NULL, sp = NULL)
}
\arguments{
\item{dat}{Input dataset}

\item{wttype}{Type of statistical weighting method; 'equal' gives equal weight to all rows, 'area' weights rows in invrse proportion to the number of rows per stratum; and 'catch' weights rows in proportion to the catch in the stratum divided by the number of rows in the stratum.}

\item{catch}{An optional vector of catch per stratum.}

\item{sp}{The species to be used for catch weighting.}
}
\value{
A vector indicating the statistical weight to apply to the stratum of which the row is a member.
}
\description{
The function returns a vector indicating a statistical weight to apply to the row, based on the stratum of which the row is a member.
}
