% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Mode}
\alias{Mode}
\title{Calulate mode of a vector.}
\usage{
Mode(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
Mode.
}
\description{
A helper function to calculate the mode (most common value) of a vector.
}
