% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{PCA_by_trip}
\alias{PCA_by_trip}
\title{Principal components analysis of species composition by 'trip'
Run PCA at the trip level using normalised species composition data.}
\usage{
PCA_by_trip(
  datr,
  Screeplotname = "screeplot set",
  allsp,
  clustid,
  saveplot = TRUE,
  savedir = "."
)
}
\arguments{
\item{datr}{The input dataset.}

\item{Screeplotname}{File name of the diagnostic Scree plot.}

\item{allsp}{Species codes / variable names to use in the PCA.}

\item{clustid}{Name of variable to aggregate across for trip-level clustering.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}
}
\value{
A list of two objects based on standard PCA and binomial PCA.
}
\description{
Principal components analysis of species composition by 'trip'
Run PCA at the trip level using normalised species composition data.
}
