% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.agg.slope.ratio}
\alias{plot.agg.slope.ratio}
\title{Make slope ratio plot for two sets of coefficients.}
\usage{
\method{plot}{agg.slope.ratio}(coefs1, coefs2, aggyr, opyr, titl,
  lab1 = "coefs1", lab2 = "coefs2", fname = NULL, addrate = T)
}
\arguments{
\item{coefs1}{Coefficients for model 1.}

\item{coefs2}{Coefficients for model 2.}

\item{aggyr}{Year sequence in the first glm.}

\item{opyr}{Year sequence in the 2nd glm.}

\item{titl}{Title of plot.}

\item{lab1}{Label for the first glm.}

\item{lab2}{Label for the second glm.}

\item{fname}{Filename for saving the plot.}

\item{addrate}{If TRUE, write the rate of change on the plot.}
}
\description{
Function to make slope ratio plot two sets of coefficients.
}
