% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.res}
\alias{plot.res}
\title{Make slope ratio plot for two glm summaries.}
\usage{
\method{plot}{res}(summ1, summ2, name1, name2)
}
\arguments{
\item{summ1}{Summary object for model 1.}

\item{summ2}{Summary object for model 2.}

\item{name1}{Label for the first glm.}

\item{name2}{Label for the second glm.}
}
\description{
Function to make slope ratio plot for 2 glm summaries, with break in the year 2000.
}
