% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.slope.ratio}
\alias{plot.slope.ratio}
\title{Make slope ratio plot for two sets of coefficients.}
\usage{
\method{plot}{slope.ratio}(coefs1, coefs2, yr, titl)
}
\arguments{
\item{coefs1}{Coefficients for model 1.}

\item{coefs2}{Coefficients for model 2.}

\item{titl}{Title of plot.}

\item{yrqtr}{Year sequence.}
}
\description{
Function to make slope ratio plot for two sets of coefficients. Assumes there are indices for all years. Probably not used any more.
}
