% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.slope.ratio2}
\alias{plot.slope.ratio2}
\title{Make slope ratio plot for two sets of coefficients.}
\usage{
\method{plot}{slope.ratio2}(coefs1, coefs2, yr1, yr2, titl)
}
\arguments{
\item{coefs1}{Coefficients for model 1.}

\item{coefs2}{Coefficients for model 2.}

\item{yr1}{Year-qtr sequence in the 1st glm.}

\item{yr2}{Year-qtr sequence in the 2nd glm.}

\item{titl}{Title of plot.}
}
\description{
Function to make slope ratio plot for two sets of coefficients. Assumes there are indices for all years. Probably not used any more.
}
