% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_catchmap}
\alias{plot_catchmap}
\title{Plot map of catches or other spatial variable.}
\usage{
plot_catchmap(indat, vbl, dcd, latlim = c(-40, 20), lonlim = c(20,
  130), brk = seq(0, 1, 0.05), brk2 = seq(0, 1, 0.1), ti = "",
  bgc = "grey", maptype = "world2")
}
\arguments{
\item{indat}{Input dataset to plot.}

\item{vbl}{The variable to plot spatially.}

\item{dcd}{Time identifier for selecting data.}

\item{latlim}{Latitude limits for plot.}

\item{lonlim}{Longitude limits for plot.}

\item{brk}{Transition levels for the image colours.}

\item{brk2}{Transition levels for the contour lines.}

\item{ti}{Title of the plot.}

\item{bgc}{Background color of the plot, defaults to grey.}

\item{maptype}{Set to world, world2, or EPO.}
}
\description{
Function to plot map of catches or other spatial variable.
}
