% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_catchmap2}
\alias{plot_catchmap2}
\title{Plot map of catches or other spatial variable.}
\usage{
plot_catchmap2(indat, vbl, dcd, latlim = c(-40, 20), lonlim = c(20, 130),
  ti = "")
}
\arguments{
\item{indat}{Input dataset to plot.}

\item{vbl}{The variable to plot spatially.}

\item{dcd}{Time identifier for selecting data.}

\item{latlim}{Latitude limits for plot.}

\item{lonlim}{Longitude limits for plot.}

\item{ti}{Title of the plot.}
}
\description{
Function to plot map of catches or other spatial variable. This version has no defined breaks, but ensures the spatial effects are constrained within the correct 1x1 square.
}
