% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_cpuemap2}
\alias{plot_cpuemap2}
\title{Plot map of cpue or the ratio of two other spatial variables.}
\usage{
plot_cpuemap2(indat, vb1, vb2, dcd, latlim = c(-40, 40),
  lonlim = c(120, 210), ti = "", delta = 1, bgc = "grey",
  maptype = "world2")
}
\arguments{
\item{indat}{Input dataset to plot.}

\item{vb1}{The variable to plot spatially.}

\item{vb2}{The variable to plot spatially.}

\item{dcd}{Time identifier for selecting data.}

\item{latlim}{Latitude limits for plot.}

\item{lonlim}{Longitude limits for plot.}

\item{ti}{Title of the plot.}

\item{delta}{The spatial resolution of the data being used.}

\item{bgc}{Background color}

\item{maptype}{Set to world, world2 or EPO.}
}
\description{
Function to plot map of cpue or other spatial variable. This version has no defined breaks, but ensures that the spatial effects are constrained within the correct 1x1 square.
}
