% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_effects}
\alias{plot_effects}
\title{Make effects plot for a glm object.}
\usage{
plot_effects(model, indat, addmain = F, addbranch = F, addalb = F,
  addother = F, ti = "")
}
\arguments{
\item{model}{glm object to plot.}

\item{indat}{Input dataset used to create the glm.}

\item{addmain}{If TRUE, plot effects for mainline.}

\item{addbranch}{If TRUE, plot effects for branchline.}

\item{addalb}{If TRUE, plot effects for albacore.}

\item{addother}{If TRUE, plot effects for other.}

\item{ti}{Not used.}
}
\description{
Function to make effects plot for a glm object. Originally developed for Japanese data in the WCPO.
}
