% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_effects_IO}
\alias{plot_effects_IO}
\title{Make effects plot for a glm object.}
\usage{
plot_effects_IO(
  model,
  indat,
  dovess = T,
  addcl = F,
  addpca = F,
  ti = "",
  dohbf = T
)
}
\arguments{
\item{model}{glm object to plot.}

\item{indat}{Input dataset used to create the glm.}

\item{dovess}{If TRUE, plot effects for vessels.}

\item{addcl}{If TRUE, plot effects for clust.}

\item{addpca}{If TRUE, plot effects for PCA.}

\item{ti}{Title for plot.}

\item{dohbf}{If TRUE, plot effects for hbf.}
}
\description{
Function to make effects plot for a glm object. Originally developed for analyses in the Indian Ocean.
}
