% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_IO}
\alias{plot_IO}
\title{Map of the Indian Ocean.}
\usage{
plot_IO(
  plot_title = "",
  uselims = c(20, 130, -50, 25),
  sp = "YFT",
  newm = T,
  lwdm = 3,
  axes = T,
  tcol = "red",
  mapfill = TRUE,
  bgc = "lightblue"
)
}
\arguments{
\item{plot_title}{Plot title.}

\item{uselims}{Latitudes and Longitudes for the edges of the map.}

\item{sp}{Region type code for the region boundaries.}

\item{newm}{If TRUE, create a new plot, otherwise add boundaries etc to existing plot.}

\item{lwdm}{Line width for boundaries.}

\item{axes}{If TRUE, create x and y axes.}

\item{tcol}{Text colour.}

\item{mapfill}{If TRUE, fill the land area of the map.}

\item{bgc}{Background colour}
}
\description{
Function to make a map of the Indian Ocean, with regional boundaries.
}
