% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_km_deviance}
\alias{plot_km_deviance}
\title{Make kmeans deviance plot.}
\usage{
plot_km_deviance(dat, allsp, r, ti, regtype = "regY", saveplot = TRUE,
  savedir = ".")
}
\arguments{
\item{dat}{Dataset to cluster.}

\item{allsp}{species code variabl names to include in the kmeans clustering process.}

\item{r}{Region number to use as a plotting label.}

\item{ti}{Use in the plot filename.}

\item{regtype}{Not used.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}

\item{savedir}{The folder where figures are saved.}
}
\description{
Function to make kmeans deviance plot, for estimating the appropriate number of clusters.
}
