% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_explore_data_functions.R
\name{plot_mean_fishing_location}
\alias{plot_mean_fishing_location}
\title{Plot the mean fishing location by season.}
\usage{
plot_mean_fishing_location(dat)
}
\arguments{
\item{dat}{Dataset to cluster.}
}
\description{
Function to make fishing location plot.
}
