% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_slope_ratio}
\alias{plot_slope_ratio}
\title{Make slope ratio plot for two sets of coefficients.}
\usage{
plot_slope_ratio(coefs1, coefs2, yr, titl)
}
\arguments{
\item{coefs1}{Coefficients for model 1.}

\item{coefs2}{Coefficients for model 2.}

\item{yr}{Year sequence.}

\item{titl}{Title of plot.}
}
\description{
Function to make slope ratio plot for two sets of coefficients. Assumes there are indices for all years. Probably not used any more.
}
