% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{plot_spfreqyq}
\alias{plot_spfreqyq}
\title{Use this plot when deciding which species to cluster.}
\usage{
plot_spfreqyq(indat, reg_struc, splist, flag, mfr = c(5, 3))
}
\arguments{
\item{indat}{Input dataset}

\item{reg_struc}{The name of the regional structure to be used.}

\item{splist}{Vector of variable names to include in the plot.}

\item{flag}{Flag to identify the fleet, used in the filename of the plot.}

\item{mfr}{The par(mfrow) variable used to define the figure layout. Defaults to c(5,3).}
}
\description{
Use this plot when deciding which species to cluster.
}
