% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plot_functions.R
\name{plotdiags}
\alias{plotdiags}
\title{Diagnostic plots for Gaussian glm.}
\usage{
plotdiags(res, ti = "", ...)
}
\arguments{
\item{res}{Fitted model for diagnostics.}

\item{ti}{Titles for the plots.}
}
\description{
The function produces 2 Gaussian diagnostic plots: a frequency histogram and a QQ plot.
}
