% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plot_functions.R
\name{qqDist}
\alias{qqDist}
\title{QQ plot for Gaussian glm.}
\usage{
qqDist(x, standardise = F, add.median = F, ...)
}
\arguments{
\item{x}{Residuals or standardized residuals.}

\item{add.median}{Add median line if TRUE.}

\item{standardize}{Standardize the residuals if TRUE.}
}
\description{
The function produces a QQ plot for a GLM.
}
