% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{resplot_all_by_yq_cl}
\alias{resplot_all_by_yq_cl}
\title{Plot all residuals by cluster and year-quarter.}
\usage{
resplot_all_by_yq_cl(a, b, mf, outdir, fname, modtype)
}
\arguments{
\item{a}{The model data.}

\item{b}{The model residuals.}

\item{mf}{The mfrow parameter, to set up the figure.}

\item{outdir}{Directory to save the results in.}

\item{fname}{File identifier, for the file name.}

\item{modtype}{Either dellog or logC, to determine the filename.}
}
\description{
The function takes the residuals extracted from a model object and boxplots them by year-quarter and cluster.
}
