% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{revtrunc}
\alias{revtrunc}
\title{Calulate remainder.}
\usage{
revtrunc(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
Remainder.
}
\description{
A helper function to calculate the remainder, x - floor(x).
}
