% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{run_standardization}
\alias{run_standardization}
\title{Run standardization models.}
\usage{
run_standardization(runpars, doflags, regstr, maxyr, do_early, stdlabs,
  twlimit = 2005)
}
\arguments{
\item{runpars}{The 'runpars' object with all variables.}

\item{doflags}{Include HBF in the models.}

\item{regstr}{The name of the regional structure to be run.}

\item{maxyr}{Include no times after the start of this year.}

\item{do_early}{If TRUE, run the additional pre-1979 and post-1979 standardizations.}
}
\description{
The function runs delta lognormal models.
}
