% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_strat_wts_functions.R
\name{samp_data}
\alias{samp_data}
\title{Randomly sample dataset.}
\usage{
samp_data(dat, n, nsamp)
}
\arguments{
\item{dat}{Input dataset}

\item{n}{A vector indicating the number of samples in the stratum of which the row is a member.}

\item{nsamp}{An integer indicating the maximum preferred number of samples per stratum.}
}
\value{
Modified dataset.
}
\description{
The function reduces the size of the dataset per stratum to approximately nsamp.
}
