% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_strat_wts_functions.R
\name{samp_data2}
\alias{samp_data2}
\title{Randomly sample dataset.}
\usage{
samp_data2(dat, p)
}
\arguments{
\item{dat}{Input dataset}

\item{p}{Proportion.}
}
\value{
Modified dataset.
}
\description{
The function reduces the size of the dataset to proportion p of the original size.
}
