% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data_functions.R
\name{select_data}
\alias{select_data}
\title{Select data for the GLM.}
\usage{
select_data(indat, runreg, runsp, mt, minqtrs = 2, maxqtrs = 500,
  addmain = F, addother = F, addalb = F, fc = "both",
  bait = "no23", llstrat = 5, doboth = F, addcl = F)
}
\arguments{
\item{indat}{Input dataset}

\item{runreg}{The region label to select data for.}

\item{runsp}{Species of interest.}

\item{mt}{Model type, used to select only nonzero sets when mt is 'deltapos'.}

\item{minqtrs}{Vessels must fish in at least this many qtrs.}

\item{maxqtrs}{Vessels must fish no more than this many qtrs.}

\item{addmain}{Include mainline in the dataset if TRUE.}

\item{addother}{Include bet or yft in the dataset if TRUE, depending on the species of interest.}

\item{addalb}{Include 'alb' in the dataset if TRUE.}

\item{fc}{Select only fishing category 1 or 2. Applies only to Japanese data with these variables reported.}

\item{bait}{For Japanese data - select certain bait types.}

\item{llstrat}{Define latlong based on this stratification level defined here.}

\item{doboth}{Remove whichever of bet and yft is not the species of interest, if FALSE.}

\item{addcl}{Deprecated, not used.}
}
\value{
Modified dataset.
}
\description{
The function chooses fields and rows of data for the GLM analysis according to various criteria. This version developed for Japanese data in the Pacific.
}
