% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data_functions.R
\name{select_data_IO2}
\alias{select_data_IO2}
\title{Select joint data for the GLM, updated 2018.}
\usage{
select_data_IO2(indat, runreg, runpars, mt, vars, yrlims = NA,
  oneflag = NA)
}
\arguments{
\item{indat}{Input dataset}

\item{runreg}{The region label to select data for.}

\item{runpars}{Settings for this run, passed as a list.}

\item{mt}{Model type, used to select only nonzero sets when mt is 'deltapos'.}

\item{vars}{Default variables to include in the dataset.}

\item{yrlims}{Bounding years for the analysis. Use integer values, because yrqtrs use 0.125 to 0.875.}

\item{oneflag}{If not NA, the flag to use in the analysis. Otherwise flags JP, KR, SY, TW are used.}
}
\value{
Modified dataset.
}
\description{
The function chooses fields and rows of data for the GLM analysis according to various criteria. This version developed for selecting from combined (joint) datasets in the Indian Ocean.
}
