% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data_functions.R
\name{select_data_JPIO}
\alias{select_data_JPIO}
\title{Select JP data for the GLM.}
\usage{
select_data_JPIO(
  indat,
  runreg,
  runsp,
  mt,
  minqtrs = 2,
  maxqtrs = 500,
  minvess = 100,
  minll = 100,
  minyrqtr = 100,
  llstrat = 5,
  addcl = NA,
  addpca = NA,
  samp = NA,
  strsmp = 30
)
}
\arguments{
\item{indat}{Input dataset}

\item{runreg}{The region label to select data for.}

\item{runsp}{Species of interest.}

\item{mt}{Model type, used to select only nonzero sets when mt is 'deltapos'.}

\item{minqtrs}{Vessels must fish in at least this many qtrs.}

\item{maxqtrs}{Vessels must fish no more than this many qtrs.}

\item{minvess}{Include only vessels with at least this many sets.}

\item{minll}{Include only grid cells with at least this many sets.}

\item{minyrqtr}{Include only year-qtrs with at least this many sets.}

\item{llstrat}{Define latlong based on this stratification level defined here.}

\item{addcl}{If not NA, defines the name of the cluster variable to include in the dataset.}

\item{addpca}{Include PCA variables in the dataset if TRUE.}

\item{samp}{Unless NA, apply random subsampling with n = samp.}

\item{strsmp}{Unless NA, apply stratified sampling with n = strsmp.}
}
\value{
Modified dataset.
}
\description{
The function chooses fields and rows of data for the GLM analysis according to various criteria. This version developed for selecting from Japanese datasets in the Indian Ocean. Probably no longer used, replaced by select_data_JointIO.
}
