% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{setup_AO_regions}
\alias{setup_AO_regions}
\title{Set up Atlantic Ocean regions}
\usage{
setup_AO_regions(dat, regB = F)
}
\arguments{
\item{dat}{Input dataset}

\item{regB}{If TRUE, set up regB}
}
\value{
Modified dataset.
}
\description{
The function sets up the Atlantic Ocean regions for datasets with lat5 and lon5 variables.
}
