% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{setup_IO_regions}
\alias{setup_IO_regions}
\title{Set up Indian Ocean regions}
\usage{
setup_IO_regions(dat, regY = F, regY1 = F, regY2 = F, regB = F,
  regB1 = F, regB2 = F, regB3 = F, regA = F, regA1 = F, regA2 = F,
  regA3 = F, regA4 = F, regA5 = F)
}
\arguments{
\item{dat}{Input dataset}

\item{regY}{If TRUE, set up regY}

\item{regY1}{If TRUE, set up regY1}

\item{regY2}{If TRUE, set up regY2}

\item{regB}{If TRUE, set up regB}

\item{regB1}{If TRUE, set up regB1}

\item{regB2}{If TRUE, set up regB2}

\item{regB3}{If TRUE, set up regB3}

\item{regA}{If TRUE, set up regA}

\item{regA1}{If TRUE, set up regA1}

\item{regA2}{If TRUE, set up regA2}

\item{regA3}{If TRUE, set up regA3}

\item{regA4}{If TRUE, set up regA4}

\item{regA5}{If TRUE, set up regA5}
}
\value{
Modified dataset.
}
\description{
The function sets up the Indian Ocean regions for datasets with lat5 and lon5 variables.
}
