% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{splitvessels}
\alias{splitvessels}
\title{Select vessels based on quarters fished.}
\usage{
splitvessels(indat, period)
}
\arguments{
\item{indat}{Input dataset}

\item{period}{The number of years before the vessel id is changed.}
}
\value{
Modified dataset.
}
\description{
The function splits all vessels' time series after a specified number of years.
}
