% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_store_functions.R
\name{storesumm}
\alias{storesumm}
\title{Extract and save summary of GLM results from model object.}
\usage{
storesumm(mod, fname)
}
\arguments{
\item{mod}{The model object.}

\item{fname}{The file name to save to.}
}
\description{
The function extacts and saves the summary of the results from a glm object.
}
