% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_store_functions.R
\name{summarize_and_store_dlx}
\alias{summarize_and_store_dlx}
\title{Summarize and store GLM results for delta lognormal, with time area interactions.}
\usage{
summarize_and_store_dlx(modb1, modb2, modp1, modp2, dat, datpos, fname,
  modlab, dohbf = dohbf, addcl = addcl, keepd = TRUE, bcorr,
  cell_areas)
}
\arguments{
\item{modb1}{The model result file from the binomial glm with interactions.}

\item{modb2}{The model result file from the binomial glm without interactions.}

\item{modp1}{The model result file from the lognormal positive glm with interactions.}

\item{modp2}{The model result file from the lognormal positive glm without interactions.}

\item{dat}{The dataset that was input to the binomial glm.}

\item{datpos}{The dataset that was input to the lognormal positive glm.}

\item{fname}{File name used for saving outputs.}

\item{modlab}{Model label used in plots and for output filenames.}

\item{dohbf}{Include hbf if TRUE.}

\item{addcl}{Use clusters if TRUE.}

\item{keepd}{Remove data from model object unless TRUE.}

\item{bcorr}{Use lognormal bias-correction if TRUE.}

\item{cell_areas}{Cell areas to use when aggregating densities across space.}
}
\description{
The function summarizes and stores the results of all the components of a delta lognormal standardization, with time area interactions.
}
