% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_store_functions.R
\name{summarize_and_store_logCx}
\alias{summarize_and_store_logCx}
\title{Summarize and store GLM results for lognormal constant glm, with time area interactions.}
\usage{
summarize_and_store_logCx(
  mod1,
  mod2,
  dat,
  fname,
  modlab,
  addcl = addcl,
  dohbf = dohbf,
  keepd = TRUE,
  mn,
  bcorr = FALSE,
  cell_areas
)
}
\arguments{
\item{mod1}{The model result file from the lognormal constant glm with interactions.}

\item{mod2}{The model result file from the lognormal constant glm without interactions.}

\item{dat}{The dataset that was input to the glm.}

\item{fname}{File name used for saving outputs.}

\item{modlab}{Model label used in plots and for output filenames.}

\item{addcl}{Use clusters if TRUE.}

\item{dohbf}{Include hbf if TRUE.}

\item{keepd}{Remove data from model object unless TRUE.}

\item{mn}{Mean value used as the constant.}

\item{bcorr}{Use lognormal bias-correction if TRUE.}

\item{cell_areas}{Cell areas to use when aggregating densities across space.}
}
\description{
The function summarizes and stores the results of all the components of a lognormal constant standardization, with time area interactions.
}
