% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_functions.R
\name{boxplots_TPCA}
\alias{boxplots_TPCA}
\title{Boxplots of principal components by set characteristics.}
\usage{
boxplots_TPCA(dat, nPCA = 6, ti = "", dohbf = T, lat5 = F,
  regtype = "regY", r = r, saveplot = TRUE, savedir = ".")
}
\arguments{
\item{dat}{Input dataset}

\item{nPCA}{Number of principal components to plot.}

\item{ti}{Plot main title.}

\item{dohbf}{If TRUE boxplot by HBF.}

\item{lat5}{If TRUE use lat5 instead of lat.}

\item{regtype}{Deprecated, not used.}

\item{r}{Deprecated, not used.}

\item{saveplot}{Should be set to FALSE when working in RMarkdown.}

\item{savedir}{The directory to save the plot.}
}
\description{
This function plots boxplots of principal components by set characteristics.
}
