% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean_functions.R
\name{dataclean_TW_std}
\alias{dataclean_TW_std}
\title{Clean data.}
\usage{
dataclean_TW_std(dat1, doHBF = F, splist = c("alb", "bet", "yft",
  "ott", "swo", "mls", "bum", "blm", "otb", "skj", "sha", "oth", "sbt"))
}
\arguments{
\item{dat1}{Input dataset}

\item{doHBF}{If TRUE, remove HBF > 25.}

\item{splist}{List of species codes}
}
\value{
Modified dataset.
}
\description{
The function sets up variables and removes bad data. Originally developed for Taiwanese data in the IO.
}
