% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_KR}
\alias{dataprep_KR}
\title{Prepare Korean longline data.}
\usage{
dataprep_KR(dat, alldat = F)
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{Deprecated.}
}
\value{
Modified dataset.
}
\description{
The function prepares Korean longline data for IO analyses.
}
