% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_TW}
\alias{dataprep_TW}
\title{Prepare Taiwanese longline data.}
\usage{
dataprep_TW(dat, alldat = F, region = "IO", splist = c("alb", "bet",
  "yft", "ott", "swo", "mls", "bum", "blm", "otb", "skj", "sha", "oth",
  "pbf", "sbt"), nomlist = "noms1")
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{Not used.}

\item{region}{IO or AO.}

\item{splist}{Define the species in the dataset.}

\item{nomlist}{The list of variable names to return.}
}
\value{
Modified dataset.
}
\description{
The function prepares Taiwanese longline data for IO analyses.
}
