% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{dataprep_TW}
\alias{dataprep_TW}
\title{Prepare Taiwanese longline data.}
\usage{
dataprep_TW(dat, alldat = F)
}
\arguments{
\item{dat}{Input dataset}

\item{alldat}{Not used.}
}
\value{
Modified dataset.
}
\description{
The function prepares Taiwanese longline data for IO analyses.
}
