% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{makedmy}
\alias{makedmy}
\title{Turn YMD data into day month and year.}
\usage{
makedmy(yy, mm, dd)
}
\arguments{
\item{yy}{Year}

\item{mm}{Month}

\item{dd}{Day}
}
\value{
dmy
}
\description{
The function provides a dmy field.
}
