% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{map_EPO}
\alias{map_EPO}
\title{Map of the Pacific Ocean with EPO boundaries.}
\usage{
map_EPO(new = FALSE, latlim = c(-40, 40), lonlim = c(140, 290))
}
\arguments{
\item{new}{Make a new plot}

\item{latlim}{Latitudes for the edges of the map.}

\item{lonlim}{Longitudes for the edges of the map.}
}
\description{
Function to make a map of the Pacific Ocean, with regional boundaries.
}
