% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{prep_indices}
\alias{prep_indices}
\title{Prepare indices, save files and plots.}
\usage{
prep_indices(resdirs, reg_strs, rgl, vartypes = c("lognC", "dellog"),
  yr1 = 1952, outdirname = "outputs/", mdt_all = c("novess_allyrs",
  "boat_allyrs", "novess_5279", "vessid_79nd"), mdti_all = c(paste0(yr1,
  "-present no vessid"), paste0(yr1, "-present vessid"), paste0(yr1,
  "-1979 no vessid"), "1979-present vessid"))
}
\arguments{
\item{resdirs}{Directories to work through.}

\item{reg_strs}{Names of the regional structures to work through.}

\item{rgl}{A list of the regions to plot in each regional structure.}

\item{vartypes}{One or both of 'lognC' and 'dellog'.}

\item{yr1}{First year of the early model, for titles.}

\item{outdirname}{The default name of the folder to create for the outputs.}
}
\description{
The function takes a list of directories with output file.
It goes through them and sets up dataframes of the yera-quarterly indices.
It generates annual indices by regressing pr ~ yr + qtr.
It saves .csv files of all indices, and generates and saves index plots.
}
