% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{setup_EPO_regions}
\alias{setup_EPO_regions}
\title{Set up Eastern Pacific Ocean regions}
\usage{
setup_EPO_regions(dat, regB = F, regB1 = F)
}
\arguments{
\item{dat}{Input dataset}

\item{regB}{If TRUE, set up regB}

\item{regB1}{If TRUE, set up regB1}
}
\value{
Modified dataset.
}
\description{
The function sets up the Eastern Pacific Ocean regions for datasets with lat5 and lon5 variables.
}
