% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{setup_EPO_regions}
\alias{setup_EPO_regions}
\title{Set up Eastern Pacific Ocean regions}
\usage{
setup_EPO_regions(dat, regBall = F, regBepo = F, regBwcpo = F)
}
\arguments{
\item{dat}{Input dataset}

\item{regBall}{If TRUE, set up regBall}

\item{regBepo}{If TRUE, set up regBepo}

\item{regBwcpo}{If TRUE, set up regBwcpo}
}
\value{
Modified dataset.
}
\description{
The function sets up the Eastern Pacific Ocean regions for datasets with lat5 and lon5 variables.
}
